finish

+++ libft/libft.h
@@ -6,7 +6,7 @@
 /*   By: pde-petr <pde-petr@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/11/04 13:43:38 by pde-petr          #+#    #+#             */
-/*   Updated: 2024/12/11 16:53:57 by pde-petr         ###   ########.fr       */
+/*   Updated: 2024/12/12 03:06:56 by pde-petr         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -55,7 +55,7 @@ char				*ft_strmapi(char const *s, char (*f)(unsigned int, char));
 void				ft_striteri(char *s, void (*f)(unsigned int, char *));
 int					ft_putchar_fd(char c, int fd);
 void				ft_putendl_fd(char *s, int fd);
-int				ft_putnbr_fd(int n, int fd);
+int				ft_putnbr_fd(long int n, int fd);
 int					ft_putstr_fd(char *s, int fd);
 t_list				*ft_lstnew(void *content);
 void				ft_lstadd_front(t_list **lst, t_list *new);
